requirements.txt
db-sqlite3
flask
flask_cors
pyjwt
pyjwt[crypto]